requirements.txt
numpy==1.21.4
pandas==1.3.4
Pillow==8.4.0
scikit_learn==1.0.1
streamlit==1.1.0
